Pulling crossvalidation factory into separate file
