+++ inferelator_ng/tests/test_amusr.py
@@ -3,7 +3,7 @@ import pandas as pd
 import numpy as np
 from inferelator_ng.regression import amusr_regression
 from inferelator_ng import amusr_workflow
-from inferelator_ng.single_cell_puppeteer_workflow import create_puppet_workflow
+from inferelator_ng.crossvalidation_workflow import create_puppet_workflow
 import numpy.testing as npt
 import pandas.util.testing as pdt
